put app back on plugin file
