--- packages/grafana-ui/src/types/app.ts
@@ -1,14 +1,0 @@
-export class AppPlugin {
-  components: {
-    ConfigCtrl?: any;
-  };
-
-  pages: { [str: string]: any };
-
-  constructor(ConfigCtrl: any) {
-    this.components = {
-      ConfigCtrl: ConfigCtrl,
-    };
-    this.pages = {};
-  }
-}
