+++ packages/grafana-ui/src/types/index.ts
@@ -9,4 +9,3 @@ export * from './threshold';
 export * from './input';
 export * from './logs';
 export * from './displayValue';
-export * from './app';
