+++ packages/grafana-ui/src/types/plugin.ts
@@ -59,3 +59,18 @@ export interface PluginMetaInfo {
   updated: string;
   version: string;
 }
+
+export class AppPlugin {
+  components: {
+    ConfigCtrl?: any;
+  };
+
+  pages: { [str: string]: any };
+
+  constructor(ConfigCtrl: any) {
+    this.components = {
+      ConfigCtrl: ConfigCtrl,
+    };
+    this.pages = {};
+  }
+}
